add foto miniaturas

+++ app/Http/Controllers/TiposController.php
@@ -6,6 +6,7 @@ use App\Models\Tipo;
 use Illuminate\Http\Request;
 use Illuminate\Pagination\Paginator;
 use Illuminate\Support\Facades\Redirect;
+use Intervention\Image\Facades\Image;
 
 class TiposController extends Controller
 {
@@ -29,8 +30,29 @@ class TiposController extends Controller
     }
 
     public function store(Request $request){
+        $this->validate($request, ['image.*', 'mimes:jpeg, jpg, gif, png']);
+        $pasta = public_path('/uploads/tipos');
+        if ($request->hasFile('icone')){
+            $foto = $request->file('icone');
+            $miniatura = Image::make($foto->path());
+            $nomeArquivo = $request->file('icone')->getClientOriginalName();
+            if(!$miniatura->resize(500,
+                                   500,
+                                   function($constraint){
+                                      $constraint->aspectRatio();
+                                   }
+                                  )
+                                  ->save($pasta.'/'.$nomeArquivo)
+               ){
+                $nomeArquivo = "semfoto.jpg";
+            }
+        } else {
+            $nomeArquivo = 'semfoto.jpg';
+        }
+
         $tipo = new Tipo();
         $tipo->fill($request->all());
+        $tipo->icone = $nomeArquivo;
         if ($tipo->save()){
             $request->session()->flash('mensagem_sucesso', "Tipo salvo!");
         } else {
@@ -41,7 +63,27 @@ class TiposController extends Controller
 
     public function update(Request $request, $tipo_id){
         $tipo = Tipo::findOrFail($tipo_id);
+        $this->validate($request, ['image.*', 'mimes:jpeg, jpg, gif, png']);
+        $pasta = public_path('/uploads/tipos');
+        if ($request->hasFile('icone')){
+            $foto = $request->file('icone');
+            $miniatura = Image::make($foto->path());
+            $nomeArquivo = $request->file('icone')->getClientOriginalName();
+            if(!$miniatura->resize(500,
+                                   500,
+                                   function($constraint){
+                                      $constraint->aspectRatio();
+                                   }
+                                  )
+                                  ->save($pasta.'/'.$nomeArquivo)
+               ){
+                $nomeArquivo = "semfoto.jpg";
+            }
+        } else {
+            $nomeArquivo = $tipo->icone;
+        }
         $tipo->fill($request->all());
+        $tipo->icone = $nomeArquivo;
         if ($tipo->save()){
             $request->session()->flash('mensagem_sucesso', "Tipo alterado!");
         } else {
@@ -57,9 +99,17 @@ class TiposController extends Controller
 
     public function deletar(Request $request, $tipo_id){
         $tipo = Tipo::findOrFail($tipo_id);
-        $tipo->delete();
-        $request->session()->flash('mensagem_sucesso',
-            'Tipo removido com sucesso');
-        return Redirect::to('tipo');
+        $lOk = true;
+        if (!empty($tipo->icone)){
+            if ($tipo->icone != 'semfoto.jpg'){
+                $lOk = unlink(public_path('uploads/tipos/').$tipo->icone);
+            }
+        }
+        if ($lOk){
+            $tipo->delete();
+            $request->session()->flash('mensagem_sucesso',
+                'Tipo removido com sucesso');
+            return Redirect::to('tipo');
+        }
     }
 }
